requirements.txt
Django>=4.2
psycopg2-binary
python-dotenv
django-ninja
PyJWT